+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0-BETA31
 
 * Added helpers for Attachment syncing.
+* Fix loading native PowerSync extension for Java targets.
 
 ## 1.0.0-BETA30